test/inspector/debugger/wasm-remove-breakpoint-expected.txt

Tests remove breakpoint from wasm scripts.

Running test: Test
Calling instantiate function.
Waiting for wasm scripts to be parsed.
Ignoring script with url v8://test/instantiate
Got wasm script: wasm://wasm/bfa7ac2d
Setting breakpoint on line 3 of wasm function
{
    columnNumber : 39
    lineNumber : 0
    scriptId : <scriptId>
}
paused No 1
Script wasm://wasm/bfa7ac2d byte offset 39: Wasm opcode 0x6b (kExprI32Sub)
Remove breakpoint
Debugger.resume
exports.main returned!
